Fix typo: four TC members per employer
This fixes a typo which incorrectly said the number of TC seats for
a particular employer is limited to three; it is in fact four.

The intent of this is to prevent a company from having a majority
of the available seats - with nine members total, that makes the
results in four seats.

Change-Id: I43e11141b4f54d766f137a6ece1943e30d49a023